requirements.txt
Django==6.0.4
djangorestframework==3.17.1
drf-spectacular==0.29.0
djangorestframework-simplejwt==5.5.1
dj-database-url==2.2.0
gunicorn==22.0.0
whitenoise==6.7.0
psycopg[binary]==3.2.10